fix(users_Mgmt): Fixed type import for 13 file
