+++ src/features/admin/UsersManagement/components/UsersTable.tsx
@@ -1,0 +1,186 @@
+import React from 'react';
+import {  EyeIcon, PencilIcon, MagnifyingGlassIcon } from '@heroicons/react/24/outline';
+import ActionDropdown from './ActionDropdown';
+import type { User } from '../types/User';
+import { formatDate } from '../utils/formatDate';
+
+interface UsersTableProps {
+  users: User[];
+  search: string;
+  setSearch: (value: string) => void;
+  roleFilter: string;
+  setRoleFilter: (value: string) => void;
+  statusFilter: string;
+  setStatusFilter: (value: string) => void;
+  isMobile: boolean;
+  paginatedUsers: User[];
+  handleClearFilters: () => void;
+  handleViewUser: (user: User) => void;
+  handleEditUser: (user: User) => void;
+  handleDeleteUser: (id: number) => void;
+}
+
+const UsersTable: React.FC<UsersTableProps> = ({
+  users,
+  search,
+  setSearch,
+  roleFilter,
+  setRoleFilter,
+  statusFilter,
+  setStatusFilter,
+  isMobile,
+  paginatedUsers,
+  handleClearFilters,
+  handleViewUser,
+  handleEditUser,
+  handleDeleteUser,
+}) => {
+  return (
+    <>
+      <div className="space-y-4 mb-6">
+        <div className="relative">
+          <MagnifyingGlassIcon className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 w-4 h-4" />
+          <input
+            type="text"
+            placeholder="Tìm kiếm theo tên hoặc email..."
+            className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+            value={search}
+            onChange={(e) => setSearch(e.target.value)}
+          />
+        </div>
+        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
+          <select
+            className="px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+            value={roleFilter}
+            onChange={(e) => setRoleFilter(e.target.value)}
+          >
+            <option value="">Tất cả vai trò</option>
+            <option value="Admin">Admin</option>
+            <option value="Judge">Judge</option>
+          </select>
+          <select
+            className="px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
+            value={statusFilter}
+            onChange={(e) => setStatusFilter(e.target.value)}
+          >
+            <option value="">Tất cả trạng thái</option>
+            <option value="active">Hoạt động</option>
+            <option value="inactive">Không hoạt động</option>
+          </select>
+          <button
+            onClick={handleClearFilters}
+            className="px-3 py-2 text-gray-600 border border-gray-300 rounded-md hover:bg-gray-50"
+          >
+            Xóa bộ lọc
+          </button>
+        </div>
+      </div>
+
+      {users.length === 0 ? (
+        <div className="text-center py-8 sm:py-12">
+          <div className="mx-auto h-12 w-12 text-gray-400">
+            <svg fill="none" viewBox="0 0 24 24" stroke="currentColor">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197M13 7a4 4 0 11-8 0 4 4 0 018 0z" />
+            </svg>
+          </div>
+          <h3 className="mt-2 text-sm sm:text-base font-medium text-gray-900">Không có người dùng</h3>
+          <p className="mt-1 text-xs sm:text-sm text-gray-500">Bắt đầu bằng cách thêm người dùng mới.</p>
+        </div>
+      ) : isMobile ? (
+        <div className="space-y-4">
+          {paginatedUsers.map((user) => (
+            <div key={user.id} className="bg-white p-4 rounded-lg shadow border">
+              <div className="flex justify-between items-start mb-2">
+                <div>
+                  <h3 className="font-semibold text-gray-900">{user.username}</h3>
+                  <p className="text-sm text-gray-600">{user.email}</p>
+                </div>
+                <ActionDropdown
+                  user={user}
+                  onView={handleViewUser}
+                  onEdit={handleEditUser}
+                  onDelete={handleDeleteUser}
+                />
+              </div>
+              <div className="flex flex-wrap gap-2 text-xs">
+                <span className={`px-2 py-1 rounded-full ${user.role === 'Admin' ? 'bg-purple-100 text-purple-800' : 'bg-blue-100 text-blue-800'}`}>
+                  {user.role}
+                </span>
+                <span className={`px-2 py-1 rounded-full ${user.isActive ? 'bg-green-100 text-green-800' : 'bg-red-100 text-red-800'}`}>
+                  {user.isActive ? 'Hoạt động' : 'Không hoạt động'}
+                </span>
+              </div>
+            </div>
+          ))}
+        </div>
+      ) : (
+        <div className="bg-white rounded-lg shadow overflow-hidden">
+          <div className="overflow-x-auto">
+            <table className="min-w-full divide-y divide-gray-200">
+              <thead className="bg-gray-50">
+                <tr>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">ID</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Tên đăng nhập</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Email</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Vai trò</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Trạng thái</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Ngày tạo</th>
+                  <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Thao tác</th>
+                </tr>
+              </thead>
+              <tbody className="bg-white divide-y divide-gray-200">
+                {paginatedUsers.map((user) => (
+                  <tr key={user.id} className="hover:bg-gray-50">
+                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">{user.id}</td>
+                    <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{user.username}</td>
+                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">{user.email}</td>
+                    <td className="px-6 py-4 whitespace-nowrap">
+                      <span className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${user.role === 'Admin' ? 'bg-purple-100 text-purple-800' : 'bg-blue-100 text-blue-800'}`}>
+                        {user.role}
+                      </span>
+                    </td>
+                    <td className="px-6 py-4 whitespace-nowrap">
+                      <span className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${user.isActive ? 'bg-green-100 text-green-800' : 'bg-red-100 text-red-800'}`}>
+                        {user.isActive ? 'Hoạt động' : 'Không hoạt động'}
+                      </span>
+                    </td>
+                    <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">{formatDate(user.createdAt)}</td>
+                    <td className="px-6 py-4 whitespace-nowrap text-sm font-medium">
+                      <div className="flex space-x-2">
+                        <button
+                          onClick={() => handleViewUser(user)}
+                          className="p-2 sm:p-1 text-blue-600 hover:text-blue-900 transition-colors"
+                          title="Xem chi tiết"
+                        >
+                          <EyeIcon className="w-5 h-5 sm:w-4 sm:h-4" />
+                        </button>
+                        <button
+                          onClick={() => handleEditUser(user)}
+                          className="p-2 sm:p-1 text-green-600 hover:text-green-900 transition-colors"
+                          title="Chỉnh sửa"
+                        >
+                          <PencilIcon className="w-5 h-5 sm:w-4 sm:h-4" />
+                        </button>
+                        <button
+                          onClick={() => handleDeleteUser(user.id)}
+                          className="p-2 sm:p-1 text-red-600 hover:text-red-900 transition-colors"
+                          title="Xóa"
+                        >
+                          <svg className="w-5 h-5 sm:w-4 sm:h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+                          </svg>
+                        </button>
+                      </div>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
+          </div>
+        </div>
+      )}
+    </>
+  );
+};
+
+export default UsersTable;
