+++ src/features/admin/UsersManagement/hooks/useResponsive.ts
@@ -1,0 +1,16 @@
+import { useState, useEffect } from 'react';
+
+export const useResponsive = () => {
+  const [isMobile, setIsMobile] = useState(false);
+
+  useEffect(() => {
+    const checkIsMobile = () => {
+      setIsMobile(window.innerWidth < 768);
+    };
+    checkIsMobile();
+    window.addEventListener('resize', checkIsMobile);
+    return () => window.removeEventListener('resize', checkIsMobile);
+  }, []);
+
+  return { isMobile };
+};
