+++ src/features/admin/UsersManagement/utils/formatDate.ts
@@ -1,0 +1,3 @@
+export const formatDate = (dateString: string) => {
+  return new Date(dateString).toLocaleString('vi-VN');
+};
